+++ extensions/obsidian/src/components/NoteListPinned.tsx
@@ -1,22 +1,23 @@
 import { Action, getPreferenceValues, Icon, Color } from "@raycast/api";
 import React, { useState, useEffect, useMemo } from "react";
 
-import { Note, SearchNotePreferences, Vault } from "../utils/interfaces";
+import { Note, SearchArguments, SearchNotePreferences, Vault } from "../utils/interfaces";
 import { NoteList } from "./NoteList";
 import { getPinnedNotes, migratePinnedNotes, resetPinnedNotes } from "../utils/pinNoteUtils";
-import { filterNotes, getListOfTags } from "../utils/utils";
+import { getListOfTags } from "../utils/utils";
+import { filterNotes } from "../utils/search";
 import { MAX_RENDERED_NOTES, NoteAction } from "../utils/constants";
 import { NoteActions, OpenNoteActions } from "../utils/actions";
 
-export function NoteListPinned(props: { vault: Vault; showTitle: boolean }) {
+export function NoteListPinned(props: { vault: Vault; showTitle: boolean; searchArguments: SearchArguments }) {
   const { searchContent } = getPreferenceValues<SearchNotePreferences>();
-  const { showTitle, vault } = props;
+  const { showTitle, vault, searchArguments } = props;
 
   migratePinnedNotes();
 
   const [pinnedNotes, setPinnedNotes] = useState<Note[]>([]);
   const [allNotes, setAllNotes] = useState<Note[]>([]);
-  const [input, setInput] = useState<string>("");
+  const [input, setInput] = useState<string>(searchArguments.searchArgument || "");
   const list = useMemo(() => filterNotes(pinnedNotes, input, searchContent), [pinnedNotes, input]);
 
   function onDelete(note: Note) {
@@ -27,7 +28,7 @@ export function NoteListPinned(props: { vault: Vault; showTitle: boolean }) {
     return (
       <Action
         title="Reset Pinned Notes"
-        icon={{ source: Icon.XmarkCircle, tintColor: Color.Red }}
+        icon={{ source: Icon.XMarkCircle, tintColor: Color.Red }}
         shortcut={{ modifiers: ["opt"], key: "r" }}
         onAction={async () => {
           if (await resetPinnedNotes(vault)) {
@@ -65,6 +66,7 @@ export function NoteListPinned(props: { vault: Vault; showTitle: boolean }) {
       tags={tags}
       isLoading={pinnedNotes === undefined}
       vault={vault}
+      searchArguments={searchArguments}
       action={actions}
       onSearchChange={setInput}
       onDelete={onDelete}